Move CNetAddr and CService to netaddress.{cpp,h}
- for CService, remove Init() and set port to 0 directly in constructor list
- make IPV6_ADDR_SCOPE defines and place in netaddress.h
- tidy/amend includes to account for the move
