+++ src/httpserver.cpp
@@ -8,7 +8,7 @@
 #include "chainparamsbase.h"
 #include "compat.h"
 #include "util.h"
-#include "netaddress.h"
+#include "netbase.h"
 #include "rpc/protocol.h" // For HTTP status codes
 #include "sync.h"
 #include "ui_interface.h"
